+++ src/components/common/modals/ModalBody.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+interface Props {
+  children: ReactNode;
+}
+
+const ModalBody = ({ children }: Props) => {
+  return <div>{children}</div>;
+};
+
+export default ModalBody;
